Add GooglePay mini-cart support Refactor GooglePay
